rename status mod
