+++ src/block/decoder.rs
@@ -2,7 +2,7 @@ use super::Block;
 use crate::asid::Asid;
 use crate::error::{ensure_ptok, extract_pterr, extract_status_or_pterr, PtError, PtErrorCode};
 use crate::event::Event;
-use crate::flags::Status;
+use crate::status::Status;
 use crate::image::Image;
 
 use crate::{EncoderDecoderBuilder, PtEncoderDecoder};
